+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed a race condition in the layerSwitcherInSidePanel behavior.
+
 ## [v2.0.1] - 2021-09-08
 
 ### Fixed